Refactor to hooks path, add rest param passthrough
