+++ client/state/sites/hooks/index.js
@@ -1,0 +1,1 @@
+export { default as useSelectedSiteSelector } from './use-selected-site-selector';
