+++ client/state/sites/hooks/test/use-selected-site-selector.js
@@ -1,0 +1,49 @@
+import { FEATURE_WP_SUBDOMAIN } from '@automattic/calypso-products';
+import * as redux from 'react-redux';
+import hasAvailableSiteFeature from 'calypso/state/selectors/has-available-site-feature';
+import { useSelectedSiteSelector } from '../';
+import { isJetpackSite } from '../../selectors';
+
+const useSelector = jest.spyOn( redux, 'useSelector' );
+
+describe( 'useSelectedSiteSelector()', () => {
+	test( 'returns selector with selected site added on call', () => {
+		const state = {
+			sites: {
+				items: {
+					123: {
+						jetpack: true,
+					},
+				},
+			},
+			ui: {
+				selectedSiteId: 123,
+			},
+		};
+		useSelector.mockImplementation( ( selector ) => selector( state ) );
+		expect( useSelectedSiteSelector( isJetpackSite ) ).toEqual( true );
+	} );
+
+	test( 'passes through rest params', () => {
+		const state = {
+			sites: {
+				features: {
+					123: {
+						data: {
+							available: {
+								[ FEATURE_WP_SUBDOMAIN ]: true,
+							},
+						},
+					},
+				},
+			},
+			ui: {
+				selectedSiteId: 123,
+			},
+		};
+		useSelector.mockImplementation( ( selector ) => selector( state ) );
+		expect( useSelectedSiteSelector( hasAvailableSiteFeature, FEATURE_WP_SUBDOMAIN ) ).toEqual(
+			true
+		);
+	} );
+} );
